rename bigquery_data_access_provider to bigquery_provider and update refs
Ruff
